+++ src/ServerFactory.php
@@ -3,10 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\Server\Status;
 
-use Innmind\Server\Status\Servers\{
-    OSX,
-    Linux,
-};
 use Innmind\Server\Control\Server as Control;
 use Innmind\TimeContinuum\Clock;
 
@@ -18,8 +14,8 @@ final class ServerFactory
         EnvironmentPath $path,
     ): Server {
         return match (\PHP_OS) {
-            'Darwin' => OSX::of($clock, $control, $path),
-            'Linux' => Linux::of($clock, $control),
+            'Darwin' => Server::osx($clock, $control, $path),
+            'Linux' => Server::linux($clock, $control),
             default => throw new \LogicException('Unsupported operating system '.\PHP_OS),
         };
     }